добавляет роутер и исправляет замечания
