+++ js/controllers/back-button-handler.js
@@ -1,0 +1,10 @@
+import render from '../router/render';
+
+/**
+ * Возвращает на экран приветствия
+ * @param {MouseEvent} evt
+ */
+export default (evt) => {
+  evt.preventDefault();
+  render(`greeting`);
+};
