+++ js/controllers/rules.js
@@ -1,23 +1,13 @@
-import gameOne from './game-1';
 import getElement from './get-element';
-import greating from './greating';
-import render from '../views/render';
-import rulesTemplate from '../models/rules';
+import onBackButtonClicked from './back-button-handler';
+import render from '../router/render';
+import rulesTemplate from '../views/rules';
 
 const rules = getElement(rulesTemplate);
 const form = rules.querySelector(`.rules__form`);
 const nameField = form.querySelector(`.rules__input`);
 const submitButton = form.querySelector(`.rules__button.continue`);
 const backButton = rules.querySelector(`.back`);
-
-/**
- * Возвращает на экран приветствия
- * @param {MouseEvent} evt
- */
-const onBackButtonClicked = (evt) => {
-  evt.preventDefault();
-  render(greating);
-};
 
 /**
  * Отключает блокировку кнопки после ввода в поле любого значения
@@ -34,7 +24,7 @@ const onNameFieldInput = () => {
  */
 const onFormSubmit = (evt) => {
   evt.preventDefault();
-  render(gameOne);
+  render(`gameOne`);
 };
 
 backButton.addEventListener(`click`, onBackButtonClicked);
