+++ js/controllers/stats.js
@@ -1,19 +1,9 @@
 import getElement from './get-element';
-import greating from './greating';
-import render from '../views/render';
-import statsTemplate from '../models/stats';
+import onBackButtonClicked from './back-button-handler';
+import statsTemplate from '../views/stats';
 
 const stats = getElement(statsTemplate);
 const backButton = stats.querySelector(`.back`);
-
-/**
- * Возвращает на экран приветствия
- * @param {MouseEvent} evt
- */
-const onBackButtonClicked = (evt) => {
-  evt.preventDefault();
-  render(greating);
-};
 
 backButton.addEventListener(`click`, onBackButtonClicked);
