+++ js/views/greeting.js
