+++ js/views/stats.js
